Header Overhaul & Store Basic Page (Unreleased)
The design of the header has been greatly altered, overall is much cleaner with a new menu panel instead of traditional buttons for better performance on mobile browsers. Added a basic, bare bones version of the store page, much work is still needed, so this store page is mainly for design purposes until functionality is added. (minecraft as a placeholder)